+++ Type/CollectionType.php
@@ -39,6 +39,11 @@ final class CollectionType extends Type implements WrappingTypeInterface
         }
 
         if ($this->isList()) {
+            if (!$type->isIdentifiedBy(TypeIdentifier::ARRAY)) {
+                trigger_deprecation('symfony/type-info', '7.3', 'Creating a "%s" that is a list and not an array is deprecated and will throw a "%s" in 8.0.', self::class, InvalidArgumentException::class);
+                // throw new InvalidArgumentException(\sprintf('Cannot create a "%s" as list when type is not "array".', self::class));
+            }
+
             $keyType = $this->getCollectionKeyType();
 
             if (!$keyType instanceof BuiltinType || TypeIdentifier::INT !== $keyType->getTypeIdentifier()) {